+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.3
 
 - Added support of va_list parameters `...` for function parameters
+- Initial support of typedef declarations
 
 ## 0.0.2